refactor : 랭킹페이지 api 분리
